refactor: extract MCP validation to shared module
Address Copilot feedback about code duplication:
- Create shared validation::is_valid_mcp_tool_format() function
- Remove duplicate MCP validation from agent.rs and skill/mod.rs
- Both validators now use the shared implementation
- Fix CC-SK-008 suggestion to mention MCP tools in all locale files

This ensures MCP format rules stay consistent across all validators
and reduces maintenance burden when validation logic changes.
